New fields in Record; added UI for Final Manifest
This week new fields were added to NPR records that the CRC sends.  We
need to include these fields and send them on to others who will use
them, so we will parse the fields and pass them through.

I've added a manifest template for the CRC to use that includes the new
fields and also has a worksheet with instructions and example data.  I
am working to have them standardize on the new template for all the
manifests that they create.

The basic UI for final manifests has been added.  It does not do
anything yet (empty workflow), but all of the UI is in place.  I
refactored the PreManifest UI slightly to take advantage of the shared
functionality between the PreManifest UI and the Final Manifest UI.

Next step is to implement the final manifest workflow.