Fix Record.compareTo bug; upgrade to POI 3.8 b4
Fixed a bug in Record.compareTo method that was causing
PrioritizedRecords to lose Records.  Only Record.equal should nearly
match compareTo() == 0 now.  Upgraded to POI 3.8 beta 4 since it claims
to fix a bug in cloning CellStyles across workbooks.  Previously it
would copy some of the style (Fonts, but not borders and colors), but
leave the workbook styles in an inconsistent state.  Now it does not
corrupt the workbook styles, but does not copy any of the style. :(

Perhaps I will see if fixing the Apache POI bug is not too hard to do.